refactor(settings): fold Actions into the Project page and move merge method to Source Control
Actions belong to a project, so a top-level Actions category was a page
for a setting that never applies globally; it is now a section on the
Project page, using the same layers icon and reset as every other
project-scoped row. The default merge method moves to Source Control as
the new server-side pullRequestMergeMethod, with the environment default
overridable per project like everything else; the pull request panel reads
it through the shared resolver. Per-checkout sidebar grouping leaves the
Project page: the device-wide toggle already lives on General, and the
override never fit the project axis. The Project page keeps name, icon,
actions, checkouts and removal.

Select placeholders render the Mixed label through the value function so
the trigger is never blank, text generation and diagnostics search items
stop demanding a single environment, and the Diagnostics link on General
renders at every scope.
